requirements.txt
# Core numerical stack
numpy==1.19.5

# TensorFlow (GPU optional)
tensorflow==2.4.4
tensorflow-gpu==2.4.4
tensorflow-text==2.4.1
tensorflow-hub==0.12.0

# NLP utilities
sentencepiece==0.1.97
indic-nlp-library==0.81
IL-Tokenizer==0.0.2
Morfessor==2.0.6

# Data handling
pandas==1.1.5
PyYAML==6.0

# HTTP (model download / optional APIs)
requests==2.27.1

# Misc
six==1.15.0
typing-extensions==3.7.4.3

